giao diện địa chỉ

+++ LapStore/app/src/main/java/com/example/lapstore/views/ProductDetail_Screen.kt
@@ -107,7 +107,7 @@ fun ProductDetail_Screen(
     LaunchedEffect(id) {
         if (id.isNotEmpty()) {
             viewModel.getSanPhamById(id)
-            hinhAnhViewModel.getHinhAnhTheoSanPham(id)
+            hinhAnhViewModel.getHinhAnhTheoSanPham(id.toInt())
         }
     }
     SideEffect {
